pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.95.0",
    "openai>=2.31.0",
    "pydantic>=2.13.0",
    "pyyaml>=6.0.3",
]
